+++ gapic-google-cloud-storage-v2/src/test/java/com/google/storage/v2/StorageClientTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Google LLC
+ * Copyright 2024 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -1382,6 +1382,8 @@ public class StorageClientTest {
             .setOwner(Owner.newBuilder().build())
             .setCustomerEncryption(CustomerEncryption.newBuilder().build())
             .setCustomTime(Timestamp.newBuilder().build())
+            .setSoftDeleteTime(Timestamp.newBuilder().build())
+            .setHardDeleteTime(Timestamp.newBuilder().build())
             .build();
     mockStorage.addResponse(expectedResponse);
 
@@ -1636,6 +1638,8 @@ public class StorageClientTest {
             .setOwner(Owner.newBuilder().build())
             .setCustomerEncryption(CustomerEncryption.newBuilder().build())
             .setCustomTime(Timestamp.newBuilder().build())
+            .setSoftDeleteTime(Timestamp.newBuilder().build())
+            .setHardDeleteTime(Timestamp.newBuilder().build())
             .build();
     mockStorage.addResponse(expectedResponse);
 
@@ -1708,6 +1712,8 @@ public class StorageClientTest {
             .setOwner(Owner.newBuilder().build())
             .setCustomerEncryption(CustomerEncryption.newBuilder().build())
             .setCustomTime(Timestamp.newBuilder().build())
+            .setSoftDeleteTime(Timestamp.newBuilder().build())
+            .setHardDeleteTime(Timestamp.newBuilder().build())
             .build();
     mockStorage.addResponse(expectedResponse);
 
@@ -1817,6 +1823,8 @@ public class StorageClientTest {
             .setOwner(Owner.newBuilder().build())
             .setCustomerEncryption(CustomerEncryption.newBuilder().build())
             .setCustomTime(Timestamp.newBuilder().build())
+            .setSoftDeleteTime(Timestamp.newBuilder().build())
+            .setHardDeleteTime(Timestamp.newBuilder().build())
             .build();
     mockStorage.addResponse(expectedResponse);
 
@@ -1886,6 +1894,8 @@ public class StorageClientTest {
             .setOwner(Owner.newBuilder().build())
             .setCustomerEncryption(CustomerEncryption.newBuilder().build())
             .setCustomTime(Timestamp.newBuilder().build())
+            .setSoftDeleteTime(Timestamp.newBuilder().build())
+            .setHardDeleteTime(Timestamp.newBuilder().build())
             .build();
     mockStorage.addResponse(expectedResponse);
 
@@ -1955,6 +1965,8 @@ public class StorageClientTest {
             .setOwner(Owner.newBuilder().build())
             .setCustomerEncryption(CustomerEncryption.newBuilder().build())
             .setCustomTime(Timestamp.newBuilder().build())
+            .setSoftDeleteTime(Timestamp.newBuilder().build())
+            .setHardDeleteTime(Timestamp.newBuilder().build())
             .build();
     mockStorage.addResponse(expectedResponse);
 
@@ -2027,6 +2039,8 @@ public class StorageClientTest {
             .setOwner(Owner.newBuilder().build())
             .setCustomerEncryption(CustomerEncryption.newBuilder().build())
             .setCustomTime(Timestamp.newBuilder().build())
+            .setSoftDeleteTime(Timestamp.newBuilder().build())
+            .setHardDeleteTime(Timestamp.newBuilder().build())
             .build();
     mockStorage.addResponse(expectedResponse);
 
@@ -2170,6 +2184,8 @@ public class StorageClientTest {
             .setOwner(Owner.newBuilder().build())
             .setCustomerEncryption(CustomerEncryption.newBuilder().build())
             .setCustomTime(Timestamp.newBuilder().build())
+            .setSoftDeleteTime(Timestamp.newBuilder().build())
+            .setHardDeleteTime(Timestamp.newBuilder().build())
             .build();
     mockStorage.addResponse(expectedResponse);