+++ storage-shared-benchmarking/src/main/java/com/google/cloud/storage/benchmarking/StorageSharedBenchmarkingUtils.java
@@ -15,25 +15,36 @@
  */
 package com.google.cloud.storage.benchmarking;
 
-import com.google.cloud.storage.Blob;
+import com.google.cloud.storage.BlobInfo;
 import com.google.cloud.storage.Storage;
-import java.time.Duration;
 
 class StorageSharedBenchmarkingUtils {
   public static long SSB_SIZE_THRESHOLD_BYTES = 1048576;
   public static int DEFAULT_NUMBER_OF_READS = 3;
 
-  public static void cleanupObject(Storage storage, Blob created) {
+  public static void cleanupObject(Storage storage, BlobInfo created) {
     storage.delete(
         created.getBlobId(), Storage.BlobSourceOption.generationMatch(created.getGeneration()));
   }
 
-  public static double calculateThroughput(double size, Duration elapsedTime) {
-    double adjustedSize =
-        size >= StorageSharedBenchmarkingUtils.SSB_SIZE_THRESHOLD_BYTES
-            ? (size / 1024D) / 1024D
-            : size / 1024D;
-    double throughput = adjustedSize / (elapsedTime.toMillis() / 1000D);
-    return throughput;
+  public static CloudMonitoringResult generateCloudMonitoringResult(
+      String op, double latency, BlobInfo created, String api, int workers) {
+    CloudMonitoringResult result =
+        CloudMonitoringResult.newBuilder()
+            .setLibrary("java")
+            .setApi(api)
+            .setOp(op)
+            .setWorkers(workers)
+            .setObjectSize(created.getSize().intValue())
+            .setChunksize(created.getSize().intValue())
+            .setCrc32cEnabled(false)
+            .setMd5Enabled(false)
+            .setCpuTimeUs(-1)
+            .setBucketName(created.getBucket())
+            .setStatus("OK")
+            .setTransferSize(created.getSize().toString())
+            .setLatency(latency)
+            .build();
+    return result;
   }
 }