+++ storage-shared-benchmarking/src/main/java/com/google/cloud/storage/benchmarking/W1R3.java
@@ -16,6 +16,8 @@
 
 package com.google.cloud.storage.benchmarking;
 
+import static com.google.cloud.storage.benchmarking.StorageSharedBenchmarkingUtils.generateCloudMonitoringResult;
+
 import com.google.cloud.storage.Blob;
 import com.google.cloud.storage.BlobInfo;
 import com.google.cloud.storage.DataGenerator;
@@ -80,7 +82,7 @@ final class W1R3 implements Callable<String> {
           printResult("READ[" + i + "]", created, elapsedTimeDownload);
         }
       }
-      StorageSharedBenchmarkingUtils.cleanupObject(storage, created);
+      StorageSharedBenchmarkingUtils.cleanupObject(storage, created.asBlobInfo());
     } catch (Exception e) {
       CloudMonitoringResult result =
           CloudMonitoringResult.newBuilder()
@@ -96,7 +98,7 @@ final class W1R3 implements Callable<String> {
               .setBucketName("")
               .setStatus("FAIL")
               .setTransferSize("")
-              .setThroughput(0)
+              .setLatency(0)
               .build();
       printWriter.println(result.formatAsCustomMetric());
     }
@@ -106,33 +108,8 @@ final class W1R3 implements Callable<String> {
   private void printResult(String op, Blob created, Duration duration) {
     if (!isWarmup) {
       printWriter.println(
-          generateCloudMonitoringResult(
-                  op,
-                  StorageSharedBenchmarkingUtils.calculateThroughput(
-                      created.getSize().doubleValue(), duration),
-                  created)
+          generateCloudMonitoringResult(op, duration.toMillis(), created.asBlobInfo(), api, workers)
               .formatAsCustomMetric());
     }
-  }
-
-  private CloudMonitoringResult generateCloudMonitoringResult(
-      String op, double throughput, Blob created) {
-    CloudMonitoringResult result =
-        CloudMonitoringResult.newBuilder()
-            .setLibrary("java")
-            .setApi(api)
-            .setOp(op)
-            .setWorkers(workers)
-            .setObjectSize(created.getSize().intValue())
-            .setChunksize(created.getSize().intValue())
-            .setCrc32cEnabled(false)
-            .setMd5Enabled(false)
-            .setCpuTimeUs(-1)
-            .setBucketName(created.getBucket())
-            .setStatus("OK")
-            .setTransferSize(created.getSize().toString())
-            .setThroughput(throughput)
-            .build();
-    return result;
   }
 }